% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{adjust_pvalues}
\alias{adjust_pvalues}
\title{Adjust P-values for Multiple Testing}
\usage{
adjust_pvalues(results_list, method = "BH")
}
\arguments{
\item{results_list}{List of MvGGE_result objects}

\item{method}{Method for p-value adjustment (default: "BH" for Benjamini-Hochberg)}
}
\value{
List of results with adjusted p-values added
}
\description{
Applies multiple testing correction to a list of MvGGE results.
}
\examples{
# Simulate multiple datasets and test
results <- list()
for (i in 1:5) {
  data <- simulate_data(n = 200, seed = i)
  results[[i]] <- test_MvGGE(data$Y1, data$Y2, data$G, data$E)
}

# Apply multiple testing correction
results_adj <- adjust_pvalues(results)

}
