% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{compute_joint_continuous}
\alias{compute_joint_continuous}
\title{Compute Joint Multi-Phenotype P-Value for Continuous Phenotypes}
\usage{
compute_joint_continuous(Y, G, E, config = MvGGE_config())
}
\arguments{
\item{Y}{Matrix of continuous phenotypes (n x m) where m >= 2}

\item{G}{Vector of genotypes (length n)}

\item{E}{Vector of environmental factors (length n)}

\item{config}{MvGGE configuration object}
}
\value{
A list containing test results
}
\description{
Performs MANOVA-based joint testing for continuous phenotypes using Wilks' lambda test.
}
\keyword{internal}
