% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset for MvGGE Analysis}
\format{
A data frame with 500 rows and 6 variables:
\describe{
  \item{Y1}{Continuous phenotype 1}
  \item{Y2}{Continuous phenotype 2}
  \item{Yb1}{Binary phenotype 1 (0/1)}
  \item{Yb2}{Binary phenotype 2 (0/1)}
  \item{G}{Genotype (0/1/2, additive coding)}
  \item{E}{Environmental exposure (0/1)}
}
}
\source{
Simulated using MvGGE::simulate_data()
}
\usage{
example_data
}
\description{
A simulated dataset demonstrating the structure needed for MvGGE analysis.
}
\details{
This dataset was generated using \code{simulate_data()} with moderate effect sizes
and correlation. It can be used to test package functionality and learn the analysis workflow.
}
\examples{
# Load example data
data(example_data)

# Basic analysis
result <- test_MvGGE(example_data$Y1, example_data$Y2, 
                     example_data$G, example_data$E)
print(result)

}
\keyword{datasets}
