% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{expit}
\alias{expit}
\title{Expit (Inverse Logit) Function}
\usage{
expit(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Vector of probabilities
}
\description{
Numerically stable implementation of the expit function.
}
\keyword{internal}
