% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{GEE_mixed}
\alias{GEE_mixed}
\title{GEE Engine for Mixed Phenotypes}
\usage{
GEE_mixed(
  parameters,
  design_matrix,
  response_matrix,
  r,
  phi,
  config = MvGGE_config()
)
}
\arguments{
\item{parameters}{Parameter vector}

\item{design_matrix}{Design matrix}

\item{response_matrix}{Response matrix}

\item{r}{Correlation parameter}

\item{phi}{Scale parameter}

\item{config}{Configuration object}
}
\value{
List with GEE results
}
\description{
Core GEE computation for mixed phenotype analysis.
}
\keyword{internal}
