% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\docType{package}
\name{MvGGE-package}
\alias{MvGGE}
\alias{MvGGE-package}
\title{MvGGE: Multi-Phenotype Joint Testing Package}
\description{
This package performs joint multi-phenotype testing for genetic main effects
and gene-environment interactions. It supports continuous, binary, and mixed
phenotype combinations using appropriate statistical methods (MANOVA, GEE).
}
\details{
The package provides three main analysis types:
\itemize{
  \item Continuous-Continuous: Uses MANOVA with Wilks' lambda test
  \item Binary-Binary: Uses bivariate GEE with Wald test
  \item Mixed (Continuous-Binary): Uses mixed GEE with Wald test
}
}
\references{
Mishra, S. & Majumdar, A. (2024). MvGGE: Multi-Phenotype Joint Testing of
Main Genetic and Gene-Environment Interaction Effects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SauMStats/MvGGE_Codes}
  \item Report bugs at \url{https://github.com/SauMStats/MvGGE_Codes/issues}
}

}
\author{
Saurabh Mishra \email{saurabh.iith.stats@gmail.com}

Arunabha Majumdar \email{arun.majum@iith.ac.in}
}
\keyword{internal}
