% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{print.MvGGE_diagnostics}
\alias{print.MvGGE_diagnostics}
\title{Print Method for MvGGE Diagnostics}
\usage{
\method{print}{MvGGE_diagnostics}(x, ...)
}
\arguments{
\item{x}{MvGGE_diagnostics object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for MvGGE Diagnostics
}
