% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{print.MvGGE_simdata}
\alias{print.MvGGE_simdata}
\title{Print Method for MvGGE Simulation Data}
\usage{
\method{print}{MvGGE_simdata}(x, ...)
}
\arguments{
\item{x}{MvGGE_simdata object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for MvGGE Simulation Data
}
