% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{print.summary.MvGGE_result}
\alias{print.summary.MvGGE_result}
\title{Print Method for MvGGE Summary}
\usage{
\method{print}{summary.MvGGE_result}(x, ...)
}
\arguments{
\item{x}{summary.MvGGE_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for MvGGE Summary
}
