% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{summary.MvGGE_result}
\alias{summary.MvGGE_result}
\title{Summary Method for MvGGE Results}
\usage{
\method{summary}{MvGGE_result}(object, ...)
}
\arguments{
\item{object}{MvGGE_result object}

\item{...}{Additional arguments (unused)}
}
\value{
A summary object
}
\description{
Summary Method for MvGGE Results
}
