% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvGGE.R
\name{vcov.MvGGE_result}
\alias{vcov.MvGGE_result}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{MvGGE_result}(object, ...)
}
\arguments{
\item{object}{MvGGE_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Variance-covariance matrix
}
\description{
Extracts the variance-covariance matrix from MvGGE results.
}
